refactor: takeScreenshot as a standalone function
