+++ itest/lib/jest.js
@@ -1,13 +1,10 @@
 /* @flow */
 
-import {writeFileSync} from "fs"
-import crypto from "crypto"
-import path from "path"
-
 import {Application} from "spectron"
 
+import {takeScreenshot} from "./app"
 import {selectors} from "../../src/js/test/integration"
-import {LOG, LOGDIR} from "./log"
+import {LOG} from "./log"
 
 export const TestTimeout = 300000
 // https://jestjs.io/docs/en/troubleshooting#unresolved-promises
@@ -19,17 +16,7 @@ export const handleError = async (
   initialError: Error,
   done: *
 ) => {
-  try {
-    let image = await app.browserWindow.capturePage()
-    let filePath = path.join(
-      LOGDIR,
-      "failure-" + crypto.randomBytes(4).toString("hex") + ".png"
-    )
-    writeFileSync(filePath, image)
-    LOG.info(`wrote out screen shot to "${filePath}"`)
-  } catch (e) {
-    LOG.error("unable to take screen shot: " + e)
-  }
+  takeScreenshot(app)
   let realError = undefined
   let notificationError = undefined
   LOG.error(`handleError: Test hit exception: ${initialError.message}`)
